Raise error if the job is aborted
I am not sure why the behaviour was changed. It is mission critical.

It is essential that the pipelines raise an error if the job is cancelled by
the scheduler. A warning is not enough. It is not reasonbale to expect people to check
the logs for warnings 100% of the time.

Jobs may be cancelled by the scheduler for exceeding memory limits and in such cases
this may result in incomplete output. Downstream tasks may still run, and the
incomplete output may go undetected but might change the outputs in a meaningful way,
which ultimately may lead to incorrect interpretation of data.